+++ app/src/main/AndroidManifest.xml
@@ -15,16 +15,16 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".movie.activity.SeatSelectionActivity"
+            android:name=".movie.seat.SeatSelectionActivity"
             android:exported="false" />
         <activity
-            android:name=".movie.activity.TicketActivity"
+            android:name=".movie.ticket.TicketActivity"
             android:exported="false" />
         <activity
-            android:name=".movie.activity.MovieDetailActivity"
+            android:name=".movie.moviedetail.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".movie.activity.MainActivity"
+            android:name=".movie.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -32,7 +32,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <receiver android:name=".movie.alarm.AlarmReceiver"
+        <receiver android:name=".movie.setting.AlarmReceiver"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
