+++ app/src/main/java/woowacourse/movie/movie/ticket/TicketActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.movie.activity
+package woowacourse.movie.movie.ticket
 
 import android.os.Bundle
 import android.util.Log
@@ -6,12 +6,12 @@ import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivityTicketBinding
-import woowacourse.movie.movie.activity.SeatSelectionActivity.Companion.BOOKING_MOVIE_KEY
 import woowacourse.movie.movie.dto.movie.BookingMovieEntity
 import woowacourse.movie.movie.dto.movie.MovieDto
 import woowacourse.movie.movie.dto.seat.SeatsDto
 import woowacourse.movie.movie.dto.ticket.TicketCountDto
 import woowacourse.movie.movie.mapper.seat.mapToSeats
+import woowacourse.movie.movie.seat.SeatSelectionActivity.Companion.BOOKING_MOVIE_KEY
 import woowacourse.movie.movie.utils.getParcelableCompat
 import java.time.LocalDate
 import java.time.LocalDateTime
