+++ app/src/main/res/layout-land/activity_seat_selection.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:id="@+id/seat_layout"
     android:background="@color/seat_background"
-    tools:context=".movie.activity.SeatSelectionActivity">
+    tools:context=".movie.seat.SeatSelectionActivity">
 
     <ScrollView
         android:layout_width="match_parent"
