+++ app/src/main/res/layout-land/activity_ticket.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movie.activity.TicketActivity">
+    tools:context=".movie.ticket.TicketActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/ticket_toolbar"
