+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movie.activity.MainActivity">
+    tools:context=".movie.MainActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:layout_width="match_parent"
