+++ app/src/main/res/layout/fragment_history.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movie.fragment.HistoryFragment">
+    tools:context=".movie.history.HistoryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/history_rv"
