+++ app/src/main/res/layout/fragment_setting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movie.fragment.SettingFragment">
+    tools:context=".movie.setting.SettingFragment">
 
     <TextView
         android:id="@+id/push_alarm"
